My first comit of Projectile Motion for Chunkers
I created this in 2012 and am just loading it now. It used to run
in Chrome but apparently Chrome doesn't support Java plugins anymore.
I exported it as an Applet and that made three folders, one each for Linux,
macrosx, and windows. Not really sure how to manage all that. Each of them
conatains a copy of the source code, so good luck with that.